Refabricação de métodos para combater Code Smells - Data Class
