+++ src/main/java/org/jfree/data/xy/XIntervalDataItem.java
@@ -42,6 +42,8 @@ package org.jfree.data.xy;
 
 import org.jfree.data.ComparableObjectItem;
 
+import java.io.Serializable;
+
 /**
  * An item representing data in the form (x, x-low, x-high, y).
  */
@@ -109,5 +111,131 @@ public class XIntervalDataItem extends ComparableObjectItem {
             return Double.NaN;
         }
     }
+
+    /* ===========================================================
+     * JFreeChart : a free chart library for the Java(tm) platform
+     * ===========================================================
+     *
+     * (C) Copyright 2000-2022, by David Gilbert and Contributors.
+     *
+     * Project Info:  http://www.jfree.org/jfreechart/index.html
+     *
+     * This library is free software; you can redistribute it and/or modify it
+     * under the terms of the GNU Lesser General Public License as published by
+     * the Free Software Foundation; either version 2.1 of the License, or
+     * (at your option) any later version.
+     *
+     * This library is distributed in the hope that it will be useful, but
+     * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+     * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+     * License for more details.
+     *
+     * You should have received a copy of the GNU Lesser General Public
+     * License along with this library; if not, write to the Free Software
+     * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301,
+     * USA.
+     *
+     * [Oracle and Java are registered trademarks of Oracle and/or its affiliates.
+     * Other names may be trademarks of their respective owners.]
+     *
+     * -------------------
+     * YWithXInterval.java
+     * -------------------
+     * (C) Copyright 2006-2022, by David Gilbert.
+     *
+     * Original Author:  David Gilbert;
+     * Contributor(s):   -;
+     *
+     */
+
+    /**
+     * A y-value plus the bounds for the related x-interval.  This curious
+     * combination exists as an implementation detail, to fit into the structure
+     * of the ComparableObjectSeries class.  It would have been possible to
+     * simply reuse the {@link YInterval} class by assuming that the y-interval
+     * in fact represents the x-interval, however I decided it was better to
+     * duplicate some code in order to document the real intent.
+     */
+    static class YWithXInterval implements Serializable {
+
+        /** The y-value. */
+        private double y;
+
+        /** The lower bound of the x-interval. */
+        private double xLow;
+
+        /** The upper bound of the x-interval. */
+        private double xHigh;
+
+        /**
+         * Creates a new instance of {@code YWithXInterval}.
+         *
+         * @param y  the y-value.
+         * @param xLow  the lower bound of the x-interval.
+         * @param xHigh  the upper bound of the x-interval.
+         */
+        public YWithXInterval(double y, double xLow, double xHigh) {
+            this.y = y;
+            this.xLow = xLow;
+            this.xHigh = xHigh;
+        }
+
+        /**
+         * Returns the y-value.
+         *
+         * @return The y-value.
+         */
+        public double getY() {
+            return this.y;
+        }
+
+        /**
+         * Returns the lower bound of the x-interval.
+         *
+         * @return The lower bound of the x-interval.
+         */
+        public double getXLow() {
+            return this.xLow;
+        }
+
+        /**
+         * Returns the upper bound of the x-interval.
+         *
+         * @return The upper bound of the x-interval.
+         */
+        public double getXHigh() {
+            return this.xHigh;
+        }
+
+        /**
+         * Tests this instance for equality with an arbitrary object.
+         *
+         * @param obj  the object ({@code null} permitted).
+         *
+         * @return A boolean.
+         */
+        @Override
+        public boolean equals(Object obj) {
+            if (obj == this) {
+                return true;
+            }
+            if (!(obj instanceof YWithXInterval)) {
+                return false;
+            }
+            YWithXInterval that = (YWithXInterval) obj;
+            if (this.y != that.y) {
+                return false;
+            }
+            if (this.xLow != that.xLow) {
+                return false;
+            }
+            if (this.xHigh != that.xHigh) {
+                return false;
+            }
+            return true;
+        }
+
+    }
+
 
 }
